--- a/.gitignore
+++ b/.gitignore
@@ -1,2 +1,3 @@
 *.DS_Store
 Gemfile.lock
+profiling_results/mastalk_profiling*