--- a/.travis.yml
+++ b/.travis.yml
@@ -1,6 +1,5 @@
 language: ruby
 rvm:
   - 2.1.2
-  - 1.9.3
 before_install:
   - gem install bundler