--- a/Gemfile
+++ b/Gemfile
@@ -4,4 +4,5 @@ gemspec
 
 group :development do
   gem 'ruby-prof', require: false
+  gem 'pry-byebug'
 end