--- a/Gemfile
+++ b/Gemfile
@@ -3,6 +3,8 @@ source 'https://rubygems.org'
 gemspec
 
 group :development do
-  gem 'ruby-prof', require: false
+  gem 'bluecloth'
   gem 'pry-byebug'
+  gem 'rdiscount'
+  gem 'ruby-prof', require: false
 end