--- a/Gemfile
+++ b/Gemfile
@@ -2,9 +2,7 @@ source 'https://rubygems.org'
 
 gemspec
 
-group :development do
-  gem 'bluecloth'
-  gem 'pry-byebug'
-  gem 'rdiscount'
-  gem 'ruby-prof', require: false
-end
+gem 'bluecloth'
+gem 'pry-byebug'
+gem 'rdiscount'
+gem 'ruby-prof', require: false