--- a/Gemfile
+++ b/Gemfile
@@ -2,7 +2,9 @@ source 'https://rubygems.org'
 
 gemspec
 
-gem 'bluecloth'
-gem 'pry-byebug'
-gem 'rdiscount'
-gem 'ruby-prof', require: false
+group :profile do
+  gem 'bluecloth'
+  gem 'pry-byebug'
+  gem 'rdiscount'
+  gem 'ruby-prof', require: false
+end