--- a/lib/mastalk/document.rb
+++ b/lib/mastalk/document.rb
@@ -27,30 +27,11 @@ module Mastalk
     private
 
     def preprocess(source)
-      # Attempt 2
-      #
       extensions.each do |regex, block|
         while source.match(regex)
           source.sub!(regex, block.call(Regexp.last_match.captures.first).strip)
         end
       end
-
-      # Attempt 1
-      #
-      #extensions.each do |regex, block|
-      #  require 'pry-byebug'
-      #  matches = source.to_enum(:scan, regex).map { Regexp.last_match }
-      #  matches.each do |m|
-      #    source.sub!(regex, block.call(m[1]).strip)
-      #  end
-      #end
-
-      #extensions.map do |regex, block|
-      #  if source.match(regex)
-      #    source.sub!(regex, block.call(Regexp.last_match.captures.first).strip)
-      #  end
-      #end
-      #preprocess(source) if extensions.any? {|regex, _| source.match(regex)}
       source
     end
   end