--- a/lib/mastalk/extensions.rb
+++ b/lib/mastalk/extensions.rb
@@ -26,9 +26,10 @@ module Mastalk
         next if file == '.' || file == '..'
         content = File.read(File.join(SNIPPETS_FOLDER, file))
         start, stop = args(content)
+        erb = ERB.new(remove_syntax_from(content))
         extension(start, stop) do |body|
           body_lines = body.strip.gsub(/(\n|\r)+/, "\n").split(/\n/)
-          ERB.new(remove_syntax_from(content)).result(binding)
+          erb.result(binding)
         end
       end