--- a/lib/mastalk/snippets/callout.html.erb
+++ b/lib/mastalk/snippets/callout.html.erb
@@ -2,5 +2,5 @@
 
 <div class='callout callout--tip'>
   <span class="callout__icon" aria-hidden="true">?</span>
-  <%= Mastalk::Document.new(body.strip).to_html(:auto_id => false) %>
+  <%= Mastalk::Document.new(body.strip).to_html(auto_id: false) %>
 </div>