--- a/lib/mastalk/snippets/callout_tool.html.erb
+++ b/lib/mastalk/snippets/callout_tool.html.erb
@@ -6,5 +6,5 @@
       <use xlink:href="#svg-icon--tool"></use>
     </svg>
   </span>
-  <%= Mastalk::Document.new(body.strip).to_html(:auto_id => false) %>
+  <%= Mastalk::Document.new(body.strip).to_html(auto_id: false) %>
 </div>