--- a/spec/fixtures/cms_page_content
+++ b/spec/fixtures/cms_page_content
@@ -1,0 +1,149 @@
+This guide to choosing the right bank account is for you if you’re trying to decide what type of account best fits your needs or you want to know which features of an account you should compare. It can also help if you’re thinking of switching your bank account.
+
+Choose the type of account that suits your needs
+Decide what charges and features you want to compare
+Decide how you want to deal with your bank
+Compare different accounts using comparison sites
+Is it worth switching your bank account?
+If you’re claiming Universal Credit
+Watch our video – How to open a bank account
+Choose the type of account that suits your needs
+
+Current accounts
+
+Most people use a current account with a bank or building society to manage their day-to-day money. It allows you to:
+
+Pay bills by Direct Debit or standing order
+Receive automated payments such as salary, wages or benefits
+Pay for things with a debit card and withdraw money from cashpoint machines
+Access to an overdraft, although this will need to be authorised by the bank
+^Read more about current accounts.^
+
+Packaged accounts
+
+$~callout
+
+If you’re thinking about opening a packaged account, be sure to check out how many of the additional benefits you would actually use and whether you can get them cheaper elsewhere.
+
+~$
+
+Some current accounts offer extra features for which they charge a fee (often between £10 and £15 a month). These are known as packaged accounts. Extras include:
+
+Special offers (eg preferential interest rates on overdrafts)
+Car breakdown cover
+Insurance cover (eg travel or mobile phone insurance)
+Extra services
+^Read more about packaged accounts.^
+
+Fee-free Basic bank accounts
+
+These are often used as a stepping stone to a current account. A basic bank account doesn’t normally have an overdraft facility but does allow you to:
+
+Pay bills by Direct Debit or standing order
+Receive payments such as salary, wages or benefits
+A fee-free basic bank account may be right for you if:
+
+You have a poor credit record or you’re on a low income
+You don’t need the extra things that a current account offers, for example an overdraft facility
+^Want to find out more about fee-free basic bank accounts? We’ve pulled together a collection of guides, tips and handy links to make sure you get the best deal.^
+
+Jam jar accounts
+
+These accounts are sometimes called budgeting accounts or rent accounts and are designed to help you budget. They let you divide your money into different 'pots' or 'jars'. You decide how much money goes into each pot by working out how much you need for your bills and how much is left over for spending or saving.
+
+They also allow you to:
+
+Pay bills by Direct Debit or standing order
+Receive payments such as salary, wages or benefits.
+One disadvantage of these accounts is that they charge a monthly fee
+
+A jam jar account might be right for you if:
+
+You want an account that helps you to budget
+You want to avoid charges for refused Direct Debits
+You rent a council or housing association property – in which case your landlord might pay the monthly fee for you
+^Read more about jam jar accounts.^
+
+Bank accounts for students and graduates
+
+Most banks will offer a specific student account usually with an interest-free overdraft up to an agreed amount.
+
+Banks often also offer attractive accounts to graduates to attempt to secure them as long-term customers.
+
+^Find out more about bank accounts for students and graduates.^
+
+Decide what charges and features you want to compare
+
+Fees, charges and overdraft costs
+
+Fees can vary a lot between banks and between accounts, with one of the highest fees being charged for going over your agreed overdraft limit (if you have one).
+
+If you regularly spend more than you have in your account, choose one which will give you an overdraft up to an agreed limit without charging fees and/ or with a low interest rate.
+
+Interest rates on credit balances
+
+If you’re careful about your spending and never go overdrawn, have a look at the accounts that pay interest on your credit balance. However, if your earnings are on the low side and you do sometimes use your overdraft, don’t pay too much attention to the credit interest – focus more on the charges instead.
+
+Incentives
+
+Many banks offer deals to attract new customers, but be sure to check if there are strings attached. Look beyond any short-term offer and make sure that, when it ends, the account will still be the best for you. Deals include:
+
+A cash incentive
+Higher interest for a period
+A monthly credit of usually around £5.
+Decide how you want to deal with your bank
+
+Do you like dealing with a person in a branch or would you prefer the convenience of telephone or Internet banking?
+
+Not all banks provide phone, Internet, mobile banking, postal and branch services so make sure you will be able to bank how you want to.
+
+If you like going into a branch, choosing a bank you can easily get to will be the most important factor.
+
+Make sure there’s a cashpoint machine that you can use free of charge near where you live or work. Otherwise you might be charged between 75p and £10 for withdrawing money.
+
+Compare different accounts using comparison sites
+
+Comparison websites are a good starting point for anyone trying to find a current account tailored to their needs.
+
+We recommend the following websites for comparing current accounts:
+
+Go Compare{:target="_blank"} – This also allows you to use the government-backed Midata tool to securely upload your past transactions for customised current account recommendations.
+Moneyfacts{:target="_blank"}
+Money Saving Expert{:target="_blank"}
+Money Supermarket{:target="_blank"}
+Remember:
+
+Comparison websites won’t all give you the same results, so make sure you use more than one site before making a decision.
+It is also important to do some research into the type of product and features you need before making a purchase or changing supplier.
+^Find out more in our guide to comparison sites.^
+
+Is it worth switching your bank account?
+
+If you’re unhappy with the service you are getting from your current bank then it’s easy to change. Your new bank will do the work for you and there is no need to deal with your old bank.
+
+It’s worth checking every year that your bank is offering everything you need. If not, think about switching.
+
+It takes just seven working days to switch your account under the Current Account Switch Service. You choose the date you want to switch, and agree this with your new bank. They will arrange for all your incoming and outgoing payments to be moved to your new account.
+
+^Find everything you need to know about switching bank account on the Current Account Switch Service website{:target="_blank"}.^
+
+The government is working with the banks on plans to give customers their account data in a simple, standardised format that can be used in comparison sites.
+
+This will make it much easier for you to do quick and easy comparisons of current accounts, including the fees, charges and benefits, and make informed decisions on whether there are better accounts for you to switch to.
+
+If you’re claiming Universal Credit
+
+If you’re getting ready for the introduction of Universal Credit, follow the link below to find out what your account options are and the pros and cons of each one.
+
+^Read more about Choosing a bank account for your Universal Credit payment.^
+
+Watch our video – How to open a bank account
+
+$~youtube_video
+
+XZoc2Cv3BAE?
+
+Video - How to open a bank account
+~$
+
+Read a transcript of this video (DOC 26.5KB)