--- a/spec/fixtures/complex_document
+++ b/spec/fixtures/complex_document
@@ -1,3 +1,0 @@
-###test\nwith markdown
-
-#### TEST\r\n\r\nwith other stuff\r\n\r\n