--- a/spec/fixtures/duplicated_snippets
+++ b/spec/fixtures/duplicated_snippets
@@ -1,0 +1,55 @@
+
+$~youtube_video
+
+XZoc2Cv3BAE?
+
+Video 1
+~$
+
+$~youtube_video
+
+qv8RpXmXrh0?
+
+Universal Credit - Online Budgeting Support
+~$
+
+$~youtube_video
+
+2eVfzMFU9X4?
+
+Top Tips To Save Hundreds On Your Food Bill
+~$
+
+$~youtube_video
+
+U9Vkn_u-W1M?
+
+How To Sell Online And Make Cash Quick
+~$
+
+$~youtube_video
+
+Q1tVzEs9zoM?
+
+How To Sell Online And Make Cash Quick 2
+~$
+
+$~callout
+
+If you’re thinking about opening a packaged account, be sure to check out how many of the additional benefits you would actually use and whether you can get them cheaper elsewhere.
+
+~$
+
+$~callout
+
+If you’re thinking about opening a packaged account, be sure to check out how many of the additional benefits you would actually use and whether you can get them cheaper elsewhere.
+
+~$
+
+$~callout
+
+If you’re buying a home with someone else, you can both take advantage of
+separate Lifetime ISAs.
+
+~$
+