--- a/spec/fixtures/expected_conversion_duplicated_snippets
+++ b/spec/fixtures/expected_conversion_duplicated_snippets
@@ -1,0 +1,35 @@
+
+<iframe frameborder="0" height="413" width="680" src="https://www.youtube.com/embed/XZoc2Cv3BAE??rel=0" title="Video 1">
+</iframe>
+
+<iframe frameborder="0" height="413" width="680" src="https://www.youtube.com/embed/qv8RpXmXrh0??rel=0" title="Universal Credit - Online Budgeting Support">
+</iframe>
+
+<iframe frameborder="0" height="413" width="680" src="https://www.youtube.com/embed/2eVfzMFU9X4??rel=0" title="Top Tips To Save Hundreds On Your Food Bill">
+</iframe>
+
+<iframe frameborder="0" height="413" width="680" src="https://www.youtube.com/embed/U9Vkn_u-W1M??rel=0" title="How To Sell Online And Make Cash Quick">
+</iframe>
+
+<iframe frameborder="0" height="413" width="680" src="https://www.youtube.com/embed/Q1tVzEs9zoM??rel=0" title="How To Sell Online And Make Cash Quick 2">
+</iframe>
+
+<div class="callout callout--tip">
+  <span class="callout__icon" aria-hidden="true">?</span>
+  <p>If you’re thinking about opening a packaged account, be sure to check out how many of the additional benefits you would actually use and whether you can get them cheaper elsewhere.</p>
+
+</div>
+
+<div class="callout callout--tip">
+  <span class="callout__icon" aria-hidden="true">?</span>
+  <p>If you’re thinking about opening a packaged account, be sure to check out how many of the additional benefits you would actually use and whether you can get them cheaper elsewhere.</p>
+
+</div>
+
+<div class="callout callout--tip">
+  <span class="callout__icon" aria-hidden="true">?</span>
+  <p>If you’re buying a home with someone else, you can both take advantage of
+separate Lifetime ISAs.</p>
+
+</div>
+