… lib/mastalk.rb