## Context

Cms is very slow at processing documents. So this is one of the optimisation parts.

First of all I added some scripts on the bin dir. Basically what they benchmark and profile the mastalk library. 

## Profiling the actual code

Before the code was doing **unnecessary recursive loops**. 

Here it shows the profiling results:

```
ruby -Ilib bin/mastalk-profiling
```

<img width="529" alt="loops" src="https://cloud.githubusercontent.com/assets/27509/21312096/60df2636-c5e2-11e6-8cd5-561845ba0790.png">

For each time that one mastalk snippet matches on the page the algorithm was doing 21 more recursive loops (calling preprocess). 
Hypothetically, If you consider 1000 pages with 10 custom snippets will be doing a lot of unnecessary loops. I changed the implementation and it makes a small difference on production data (on my Mac was from 1 minute and 20 seconds to 50 seconds but it is still slow). 

## Benchmarking

The benchmark parses the same document 10, 100, 200, 500 and 1000 times. The document was based on one original page on CMS. **But in production data the time will be much slower since there are a lot of pages that has more snippets than the current user for the benchmarking**. So this is more like a test to identify slow spots. 

Before the PR:

```
⤷  ruby -Ilib bin/mastalk-benchmark
       user     system      total        real
Parsing 10 documents  0.210000   0.010000   0.220000 (  0.217775)
Parsing 100 documents  1.710000   0.010000   1.720000 (  2.735932)
Parsing 200 documents  3.370000   0.040000   3.410000 (  5.450446)
Parsing 500 documents  8.540000   0.060000   8.600000 (  10.644718)
Parsing 1000 documents 17.180000   0.120000  17.300000 ( 20.365242)
```

After PR:

```
⤷  ruby -Ilib bin/mastalk-benchmark
       user     system      total        real
Parsing 10 documents  0.200000   0.020000   0.220000 (  0.230090)
Parsing 100 documents  1.670000   0.020000   1.690000 (  1.695611)
Parsing 200 documents  3.340000   0.020000   3.360000 (  3.364838)
Parsing 500 documents  8.290000   0.050000   8.340000 (  8.348263)
Parsing 1000 documents 16.410000   0.100000  16.510000 ( 16.545010)
```

## Markdown library

I created a script to measure different markdown libraries and it shows that **kramdown is very slow to our case in comparison to the other libraries**. The benchmark parses the CMS page 1000 times:

```
⤷  ruby -Ilib bin/markdown-libraries-benchmark
                      user     system      total        real
Kramdown          9.250000   0.050000   9.300000 (  9.318847)
Bluecloth         0.660000   0.010000   0.670000 (  0.679713)
RDiscount         0.290000   0.010000   0.300000 (  0.300225)
```

I tried to change to Rdiscount or BlueCloth But the problem that the output of the libraries is a little different and makes our specs fail. 

**So since I am adding a cache on CMS pages I stopped just with this small optimisations.**

Conclusion: **It is still slow but the kramdown library is taking a long time too.**

I will squash the commits after the approval.