@Tr4pSt3R 2016-12-19
How does it speed up the requests, please @tomas-stefano? I'm curious about picking up new performance enhancement tricks.