@tomas-stefano 2016-12-19
As we discussed, the core is the Mastalk::Document#preprocess method that was doing unnecessary loops and regex matching the number of snippets multiplying by the occurrences of the snippet of the page and multiply by the number of pages to be processed.