@Tr4pSt3R 2016-12-19
Yeah, thanks for the explanation earlier on. I was having a hard time wrapping my mind around the whole thing but I understand the motive behind the changes a lot better now. 
Good work @tomas-stefano